Starting to add the image viewing flow to the conversationlist coordinator.
