+++ iOS/Flows/Conversation/Conversation List/ConversationListCoordinator+Presentation.swift
@@ -16,12 +16,11 @@ extension ConversationListCoordinator {
         guard let query = Circle.query() else { return }
         query.whereKey("owner", equalTo: User.current()!)
         query.getFirstObjectInBackground { [unowned self] object, error in
-            if let circle = object as? Circle {
-                let coordinator = CircleCoordinator(with: circle,
-                                                    router: self.router,
-                                                    deepLink: self.deepLink)
-                self.present(coordinator)
-            }
+            guard let circle = object as? Circle else { return }
+            let coordinator = CircleCoordinator(with: circle,
+                                                router: self.router,
+                                                deepLink: self.deepLink)
+            self.present(coordinator)
         }
     }
 
@@ -224,14 +223,38 @@ extension ConversationListCoordinator {
         self.present(coordinator, finishedHandler: nil)
     }
     
-    func presentAttachements() {
+    func presentAttachments() {
         let coordinator = AttachmentsCoordinator(router: self.router, deepLink: self.deepLink)
         
         self.present(coordinator) { [unowned self] result in
             self.handle(attachmentOption: result)
         }
     }
-    
+
+    func handle(attachmentOption option: AttachmentOption) {
+        switch option {
+        case .attachements(let array):
+            guard let first = array.first else { return }
+            let text = self.conversationListVC.messageInputController.swipeInputView.textView.text ?? ""
+            Task.onMainActorAsync {
+                guard let kind = try? await AttachmentsManager.shared.getMessageKind(for: first, body: text) else { return }
+                self.conversationListVC.messageInputController.currentMessageKind = kind
+            }
+        case .capture:
+            self.presentPhotoCapture()
+        case .audio:
+            break
+        case .giphy:
+            break
+        case .video:
+            break
+        case .library:
+            break
+            #warning("Problem when dismissing picker, and not being able to assign first responder")
+            //self.presentPhotoLibrary()
+        }
+    }
+
     func presentConversationDetail() {
         guard let cid = self.activeConversation?.cid else { return }
